requirements.txt
playwright==1.36.0
pyparsing==3.1.0
pytest==7.4.0
pytest-base-url==2.0.0
pytest-playwright==0.3.3
python-dotenv==1.0.0
allure-pytest==2.13.2